requirements.txt
Flask == 2.0.2
Flask-RESTful == 0.3.9
Flask-Cors == 3.0.10
matplotlib == 3.3.4
Pillow == 8.4.0
opencv-python == 4.5.4.58
pymongo == 3.12.1
paddleocr == 2.3.0.1
paddlepaddle==2.0.0 -i https://mirror.baidu.com/pypi/simple
